Project Site
Worked very heavily on the site.
Added an entire cookie page with table and form.
Fixed information stored in the accordion for the Events page.
Added Home/welcome page information with troop leader info and big Welcome header.
Cleaned Up all HTML and CSS validation Errors so this is a checkpoint for a site that has no errors.